requirements.txt
flask
flask-cors
requests
pymongo
python-dotenv
selenium
webdriver-manager